Refactor new types to separate file
I refactored out the new types and their trait implementations to a new
file, `src/types.rs`, because it was really long and it crowded the
implementations of more critical things, making the code hard to read.

I also reordered things within the file to lump related items together,
to hopefully make the code more readable.
